Cherry-pick 314944@main (74eb1295e28b). https://bugs.webkit.org/show_bug.cgi?id=316766
    Undefined left-shift in AV1 readULEBSize()
    https://bugs.webkit.org/show_bug.cgi?id=316766
    rdar://179211459

    Reviewed by Pascoe.

    decodedByte is a uint8_t that integer-promotes to a 32-bit int, so decodedByte << (7 * cptr) is
    undefined behavior for any shift count >= 32. Shift a size_t instead.

    * Source/WebCore/platform/graphics/AV1Utilities.cpp:
    (WebCore::readULEBSize):

    Canonical link: https://commits.webkit.org/314944@main

Canonical link: https://commits.webkit.org/305877.741@webkitglib/2.52

+++ Source/WebCore/platform/graphics/AV1Utilities.cpp
@@ -1311,7 +1311,7 @@ static size_t readULEBSize(std::span<const uint8_t> data, size_t& index)
 
         uint8_t dataByte = data[index++];
         uint8_t decodedByte = dataByte & 0x7f;
-        value |= decodedByte << (7 * cptr);
+        value |= static_cast<size_t>(decodedByte) << (7 * cptr);
         if (value >= std::numeric_limits<uint32_t>::max())
             return 0;
         if (!(dataByte & 0x80))
